flatten skill service facade into skill.ts
